yard package

+++ packages/contracts/contracts/Factory.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.11;
 
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
-import {IFactory} form "./IFactory.sol";
+import {IFactory} from "./";
 
 interface IOffscriptNFT {
     function mintPublic(
@@ -19,13 +19,13 @@ contract Factory is Ownable, IFactory {
         nft = IOffscriptNFT(_address);
     }
 
-    function mint(uint256 _optionId, address _toAddress) public {
+    function mint(uint8 _optionId, address _toAddress) public {
         if (_optionId == 0) {
         nft.mintPublic(_toAddress);
         }
     }
 
-    function tokenURI(uint256 _optionId) public view returns (string){
+    function tokenURI(uint8 _optionId) public view returns (string){
         bytes(baseURI).length > 0
                 ? string(
                     abi.encodePacked(
